+++ crates/vertigo-macro/src/include_static.rs
@@ -1,6 +1,6 @@
 use proc_macro::{Span, TokenStream};
 use quote::quote;
-use std::path::PathBuf;
+use std::{error::Error, path::PathBuf};
 
 use crate::get_target_dir::get_target_wasm_dir_with;
 
@@ -19,7 +19,7 @@ pub(crate) fn include_static_inner(input: TokenStream) -> TokenStream {
     }
 }
 
-fn bundle_file(mut file_path: PathBuf, file: String) -> Result<String, String> {
+fn bundle_file(mut file_path: PathBuf, file: String) -> Result<String, Box<dyn Error>> {
     // Remove quoting from file name
     let file_inner = &file[1..file.len() - 1];
 
@@ -30,20 +30,20 @@ fn bundle_file(mut file_path: PathBuf, file: String) -> Result<String, String> {
     let file_path = WasmPath::new(file_path);
 
     if !file_path.exists() {
-        return Err(format!("File does not exist: {}", file_path.as_string()));
+        return Err(format!("File does not exist: {}", file_path.as_string()).into());
     }
 
     if std::env::var("VERTIGO_BUNDLE").is_ok() {
         // Intermediate directory
-        let file_name = file_path.file_name();
+        let file_name = file_path.file_name()?;
         let file_static_target = get_target_wasm_dir_with(&["static", "included", &file_name]);
 
-        let file_path_content = file_path.read();
-        let hash = file_static_target.save_with_hash(file_path_content.as_slice());
+        let file_path_content = file_path.read()?;
+        let hash = file_static_target.save_with_hash(file_path_content.as_slice())?;
 
         // Final public path in the build (can be later mangled by dynamic path dispatch though)
         let Ok(public_path) = std::env::var("VERTIGO_PUBLIC_PATH") else {
-            return Err(r#"The environment variable "VERTIGO_PUBLIC_PATH" is missing"#.to_string());
+            return Err("The environment variable \"VERTIGO_PUBLIC_PATH\" is missing".into());
         };
 
         let http_path = format!("{public_path}/{hash}");