+++ crates/vertigo-macro/src/wasm_path.rs
@@ -1,6 +1,6 @@
 use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine as _};
 use crc::{Crc, CRC_64_ECMA_182};
-use std::path::PathBuf;
+use std::{error::Error, path::PathBuf};
 
 #[derive(Debug)]
 pub(crate) struct WasmPath {
@@ -16,53 +16,57 @@ impl WasmPath {
         self.path.to_string_lossy().into_owned()
     }
 
-    pub fn file_stem(&self) -> String {
-        self.path
+    pub fn file_stem(&self) -> Result<String, Box<dyn Error>> {
+        Ok(self
+            .path
             .file_stem()
-            .unwrap()
+            .ok_or("Unable to get file stem")?
             .to_string_lossy()
-            .into_owned()
+            .into_owned())
     }
 
-    pub fn file_name(&self) -> String {
-        self.path
+    pub fn file_name(&self) -> Result<String, Box<dyn Error>> {
+        Ok(self
+            .path
             .file_name()
-            .unwrap()
+            .ok_or("Unable to get file name")?
             .to_string_lossy()
-            .into_owned()
+            .into_owned())
     }
 
-    pub fn file_extension(&self) -> String {
-        self.path
+    pub fn file_extension(&self) -> Result<String, Box<dyn Error>> {
+        Ok(self
+            .path
             .extension()
-            .unwrap()
+            .ok_or("Unable to get file extension")?
             .to_string_lossy()
-            .into_owned()
+            .into_owned())
     }
 
-    pub fn save(&self, content: &[u8]) {
+    pub fn save(&self, content: &[u8]) -> Result<(), Box<dyn Error>> {
         use std::fs::File;
         use std::io::prelude::*;
 
-        let mut f = File::create(&self.path).unwrap();
-        f.write_all(content).unwrap();
+        let mut f = File::create(&self.path)?;
+        f.write_all(content)?;
+        Ok(())
     }
 
-    pub fn save_with_hash(&self, content: &[u8]) -> String {
-        let file_name = self.file_stem();
+    pub fn save_with_hash(&self, content: &[u8]) -> Result<String, Box<dyn Error>> {
+        let file_name = self.file_stem()?;
         let hash = get_hash(content);
-        let file_ext = self.file_extension();
+        let file_ext = self.file_extension()?;
 
-        let parent = self.path.parent().unwrap();
+        let parent = self.path.parent().ok_or("Unable to get parent directory")?;
         let target_file_name = format!("{file_name}.{hash}.{file_ext}");
         let target_path = Self::new(parent.join(&target_file_name));
 
-        target_path.save(content);
-        target_file_name
+        target_path.save(content)?;
+        Ok(target_file_name)
     }
 
-    pub fn read(&self) -> Vec<u8> {
-        std::fs::read(&self.path).unwrap()
+    pub fn read(&self) -> Result<Vec<u8>, std::io::Error> {
+        std::fs::read(&self.path)
     }
 
     pub fn push<P: AsRef<std::path::Path>>(&mut self, names: &[P]) {