+++ crates/vertigo/src/tests/autojsjson/rename.rs
@@ -32,7 +32,9 @@ fn test_serialize_and_deserialize_struct() {
     assert!(hash_map.contains_key("fancy data name"));
     assert!(!hash_map.contains_key("data"));
 
-    let restored_obj = TestObj::from_json(JsJsonContext::new(""), test_obj_json).unwrap();
+    let Ok(restored_obj) = TestObj::from_json(JsJsonContext::new(""), test_obj_json) else {
+        panic!("Error deserializing test_obj");
+    };
 
     assert_eq!(test_obj, restored_obj);
 }